+++ AddContacts/FeatureAPI.h
@@ -47,5 +47,5 @@ typedef void(^APICompletionBlock)(NSError *error);
 
 //Video Feature Set
 -(void)addVideoWithFileSize:(NSUInteger) fileSize withCompletionBlock:(APICompletionBlock)completionBlock;
--(void)addVideoWithDuration:(NSUInteger) duration withCompletionBlock:(APICompletionBlock)completionBlock;
+-(void)addVideoWithDuration:(int) duration withCompletionBlock:(APICompletionBlock)completionBlock;
 @end
